v5_final

+++ ft_printf.c
@@ -6,7 +6,7 @@
 /*   By: agilles <agilles@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/30 17:12:32 by agilles           #+#    #+#             */
-/*   Updated: 2023/11/08 17:00:04 by agilles          ###   ########.fr       */
+/*   Updated: 2023/11/09 16:27:21 by agilles          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ int	ft_printarg(char c, va_list *arg, int len)
 	else if (c == 's')
 		len = ft_put_s(va_arg(*arg, char *), len);
 	else if (c == 'p')
-		len = ft_put_p(va_arg(*arg, unsigned long long), len);
+		len = ft_put_p(va_arg(*arg, unsigned long), len);
 	else if (c == 'd' || c == 'i')
 		len = ft_put_di(va_arg(*arg, int), len);
 	else if (c == 'u')
